[refactor]: created exceptions folder in src, renamed docxredactor.py and exceptions_docx.py, moved docx_exceptions.py in src/exceptions/
